feat(data-validation) Add structure and validations
Adds restructing of app
Removes ui templates
Adds validation of data inputs
Refactors tests to handle validation

[Delivers #158623416]
