+++ app/api/meals.py
@@ -1,17 +1,17 @@
 """
 Module contain API resources for exposing meals
 """
-from flask_restplus import Resource, reqparse, fields
+from flask_restplus import Resource, reqparse, fields, abort
 from flask import g
 from ..models import Meal
 from .decorators import authenticate, admin_required
-from .common import str_type, price_type
+from .common import str_type
 from . import api
 
 MEAL_MODAL = api.model('Meal', {
-    'title': fields.String('Meal title'),
-    'price': fields.String('Meal price'),
-    'description': fields.String('Meal Description')
+    'title': fields.String(max_length=64),
+    'price': fields.Integer(min=100),
+    'description': fields.String(max_length=64)
 })
 
 
@@ -34,17 +34,16 @@ class MealsResource(Resource):
 
     @authenticate
     @admin_required
-    @api.expect(MEAL_MODAL)
+    @api.expect(MEAL_MODAL, validate=True)
     @api.header('Authorization', type=str, description='Authentication token')
     def post(self):
         """
         Allows a business to add new meals
         """
         parser = reqparse.RequestParser()
-        parser.add_argument('title', type=str_type,
-                            required=True, help='Title field is required')
-        parser.add_argument('price', type=price_type, required=True)
-        parser.add_argument('description', type=str)
+        parser.add_argument('title', type=str_type, required=True)
+        parser.add_argument('price', type=int, required=True)
+        parser.add_argument('description', type=str_type)
 
         args = parser.parse_args()
         user = g.current_user
@@ -69,14 +68,12 @@ class MealResource(Resource):
         meal = Meal.query.filter_by(
             catering=user.catering).filter_by(id=meal_id).first()
         if not meal:
-            return {
-                'error': 'Bad request, no meal with such id exists'
-            }, 400
+            abort(code=400, message='No meal with such id {} exists'.format(meal_id))
         return meal.to_dict(), 200
 
     @authenticate
     @admin_required
-    @api.expect(MEAL_MODAL)
+    @api.expect(MEAL_MODAL, validate=True)
     @api.doc(responses={200: 'Success', 400: 'Bad request',
                         401: 'Authorization failed'})
     @api.header('Authorization', type=str, description='Authentication token')
@@ -88,23 +85,19 @@ class MealResource(Resource):
         meal = Meal.query.filter_by(
             catering=user.catering).filter_by(id=meal_id).first()
         if not meal:
-            return {
-                'error': 'Bad request, no meal with such id exists'
-            }, 400
+            abort(code=400, message='No meal with such id {} exists'.format(meal_id))
+
         parser = reqparse.RequestParser()
         parser.add_argument('title', type=str)
-        parser.add_argument('price', type=price_type)
+        parser.add_argument('price', type=int)
         parser.add_argument('description', type=str)
         args = parser.parse_args()
 
         modified = meal.modify(args)
         if modified:
             meal.save()
-        else:
-            return {
-                'error': 'No fields specified to be modified'
-            }, 400
-        return meal.to_dict(), 200
+            return meal.to_dict(), 200
+        abort(code=400, message='Specify a field or fields to be edited')
 
     @authenticate
     @admin_required
@@ -119,8 +112,7 @@ class MealResource(Resource):
         if meal:
             meal.delete()
             return {
-                'message': 'successfully deleted'
+                'status': 'ok',
+                'message': 'meal successfully deleted'
             }, 200
-        return {
-            'error': 'no meal with such id exists'
-        }, 400
+        abort(code=400, message='No meal with such id {} exists'.format(meal_id))
