--- app/main/__init__.py
@@ -1,4 +1,0 @@
-from flask import Blueprint
-
-main = Blueprint('main', __name__, template_folder='pages')
-from .import views  # noqa
