--- app/main/views.py
@@ -1,7 +1,0 @@
-from .import main
-from flask import render_template
-
-
-@main.route('/', methods=['GET'])
-def index():
-    return render_template('index.html')
