+++ app/models/__init__.py
@@ -1,0 +1,6 @@
+from .catering import Catering
+from .user import User
+from .meal import Meal
+from .menu import Menu
+from .role import Role
+from .order import Order
