--- app/templates/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Meal App</title>
-    <style>
-        .container {
-            text-align: center;
-        }
-    </style>
-</head>
-
-<body class="container">
-    <h1>Book a meal</h1>
-    <h3>Welcome </h3>
-    <p>API Documentation can be found
-        <a href="/api/v1/#">here</a>
-    </p>
-</body>
-
-</html>
