--- ui/admin/choose_options.html
@@ -1,87 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="../css/jquery-ui.min.css">
-        <link rel="stylesheet" href="../css/main.css">
-        <link rel="stylesheet" href="../css/admin.css">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css">
-        <title>Book A meal:Admin</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html">Admin</a>
-                <a href="./mealoptions.html" class="active">Meal Options</a>
-                <a href="./menus.html">Menus</a>
-                <a href="./orders.html">Order History</a>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <main>
-            <div class="admin-container">
-                <section class="stats">
-                    <div>
-                        <button class="btn btn-g trigger">Add to Menu</button>
-                    </div>
-                </section>
-                <section class="meals">
-                    <div class="meal-o">
-                        <div class="meal-actions">
-                            <input checked type="checkbox">
-                        </div>
-                        <div class="thumbnail">
-                            <img src="../img/coffee.jpg" alt="">
-                        </div>
-                        <div class="meal-detail">
-                            <h4>Cherry Tomatoes:
-                                <span>UGX 3000</span>
-                            </h4>
-                            <label>English Coffee with cinamon flavour</label>
-                        </div>
-                    </div>
-                    <div class="meal-o">
-                        <div class="meal-actions">
-                            <input type="checkbox">
-                        </div>
-                        <div class="thumbnail">
-                            <img src="../img/oranges.jpg" alt="">
-                        </div>
-                        <div class="meal-detail">
-                            <h4>Orange Juice:
-                                <span>UGX 3000</span>
-                            </h4>
-                            <label>Lorem ipsum with cinamon flavour</label>
-                        </div>
-                    </div>
-                    <div class="meal-o">
-                        <div class="meal-actions">
-                            <input type="checkbox">
-                        </div>
-                        <div class="thumbnail">
-                            <img src="../img/coffee.jpg" alt="">
-                        </div>
-                        <div class="meal-detail">
-                            <h4>Cherry Tomatoes:
-                                <span>UGX 3000</span>
-                            </h4>
-                            <label>English Coffee with cinamon flavour</label>
-                        </div>
-                    </div>
-                </section>
-            </div>
-        </main>
-        <footer>
-
-        </footer>
-        <script src="../js/jquery.min.js"></script>
-        <script src="../js/admin.js"></script>
-        <script>
-        </script>
-    </body>
-
-</html>
