--- ui/admin/edit_mealoption.html
@@ -1,58 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="../css/jquery-ui.min.css">
-        <link rel="stylesheet" href="../css/main.css">
-        <link rel="stylesheet" href="../css/admin.css">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css">
-        <title>Book A meal:Admin</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html">Admin</a>
-                <a href="./mealoptions.html" class="active">Meal Options</a>
-                <a href="./menus.html">Menus</a>
-                <a href="./orders.html">Order History</a>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <main>
-            <div class="admin-container">
-                <section class="stats">
-                    <h4>Editing Meal Option</h4>
-                </section>
-                <section class="meals">
-                    <div>
-                        <div class="modal-content">
-                            <form method="POST" enctype="multipart/form-data">
-                                <div class="form-container">
-                                    <input class="form-input" placeholder="Meal Title" value=" Lorem ipsum connectier meal" type="text" name="title" autocomplete="off">
-                                    <input placeholder="Meal Price" class="form-input" type="number" autocomplete="off" value="30000" id="price" name="price">
-                                    <textarea name="desc" id="" placeholder="Description" cols="30" rows="10">Lorem ipsum dolor sit amet consectetur, adipisicing elit. Quaerat beatae voluptatum excepturi!
-                                        Et dignissimos sed commodi doloribus omnis numquam nisi
-                                    </textarea>
-                                    <input type="file" name="image">
-                                    <input type="button" class="btn btn-blue" value="Edit">
-                                </div>
-                            </form>
-                        </div>
-                    </div>
-                </section>
-            </div>
-        </main>
-        <footer>
-
-        </footer>
-        <script src="../js/jquery.min.js"></script>
-        <script src="../js/admin.js"></script>
-        <script>
-        </script>
-    </body>
-
-</html>
