--- ui/admin/mealoptions.html
@@ -1,132 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="../css/jquery-ui.min.css">
-        <link rel="stylesheet" href="../css/main.css">
-        <link rel="stylesheet" href="../css/admin.css">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css">
-        <title>Book A meal:Admin</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html">Admin</a>
-                <a href="./mealoptions.html" class="active">Meal Options</a>
-                <a href="./menus.html">Menus</a>
-                <a href="./orders.html">Order History</a>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <main>
-            <div class="admin-container">
-                <section class="stats">
-                    <div>
-                        <button class="btn btn-g trigger">New Meal Option</button>
-                    </div>
-                </section>
-                <section class="meals">
-                    <div class="meal-o">
-                        <div class="meal-actions">
-                            <i onclick="deleteItem()" class="fa fa-close"></i>
-                            <a href="./edit_mealoption.html">
-                                <i class="fa fa-edit"></i>
-                            </a>
-                        </div>
-                        <div class="thumbnail">
-                            <img src="../img/coffee.jpg" alt="">
-                        </div>
-                        <div class="meal-detail">
-                            <h4>Beef with rice:
-                                <span>UGX 3000</span>
-                            </h4>
-                            <label>English Coffee with cinamon flavour</label>
-                        </div>
-                    </div>
-                    <div class="meal-o">
-                        <div class="meal-actions">
-                            <i onclick="deleteItem()" class="fa fa-close"></i>
-                            <a href="./edit_mealoption.html">
-                                <i class="fa fa-edit"></i>
-                            </a>
-                        </div>
-                        <div class="thumbnail">
-                            <img src="../img/oranges.jpg" alt="">
-                        </div>
-                        <div class="meal-detail">
-                            <h4>Orange Juice:
-                                <span>UGX 3000</span>
-                            </h4>
-                            <label>Lorem ipsum with cinamonerf vour</label>
-                        </div>
-                    </div>
-                    <div class="meal-o">
-                        <div class="meal-actions">
-                            <i onclick="deleteItem()" class="fa fa-close"></i>
-                            <a href="./edit_mealoption.html">
-                                <i class="fa fa-edit"></i>
-                            </a>
-                        </div>
-                        <div class="thumbnail">
-                            <img src="../img/coffee.jpg" alt="">
-                        </div>
-                        <div class="meal-detail">
-                            <h4>Beef with fries:
-                                <span>UGX 3000</span>
-                            </h4>
-                            <label>Lorem ispum with cinamonerf vour</label>
-                        </div>
-                    </div>
-                    <div class="modal">
-                        <div class="modal-content">
-                            <span class="close-button">&times;</span>
-                            <h3>New Meal Option</h3>
-                            <form method="POST" enctype="multipart/form-data">
-                                <div class="form-container">
-                                    <input class="form-input" placeholder="Meal Title" type="text" name="title" autocomplete="off">
-                                    <input placeholder="Meal Price" class="form-input" type="number" autocomplete="off" id="price" name="price">
-                                    <textarea name="description" id="" placeholder="Description" cols="30" rows="10"></textarea>
-                                    <input type="file" name="image">
-                                    <input type="button" class="btn btn-blue" value="Add">
-                                </div>
-                            </form>
-                        </div>
-                    </div>
-                </section>
-            </div>
-        </main>
-        <footer>
-
-        </footer>
-        <script src="../js/jquery.min.js"></script>
-        <script src="../js/admin.js"></script>
-        <script>
-            $(document).ready(function () {
-
-                //trigger displaying and closing of a modal
-                var modal = document.querySelector(".modal");
-                var trigger = document.querySelector(".trigger");
-                var closeButton = document.querySelector(".close-button");
-
-                function toggleModal() {
-                    modal.classList.toggle("show-modal");
-                }
-
-                function windowOnClick(event) {
-                    if (event.target === modal) {
-                        toggleModal();
-                    }
-                }
-
-                trigger.addEventListener("click", toggleModal);
-                closeButton.addEventListener("click", toggleModal);
-                window.addEventListener("click", windowOnClick);
-            });
-        </script>
-    </body>
-
-</html>
