--- ui/admin/menus.html
@@ -1,144 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="../css/jquery-ui.min.css">
-        <link rel="stylesheet" href="../css/main.css">
-        <link rel="stylesheet" href="../css/admin.css">
-        <link rel="stylesheet" href="../css/jquery.dataTables.min.css">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css">
-        <title>Book A meal:Admin</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html">Admin</a>
-                <a href="./mealoptions.html">Meal Options</a>
-                <a href="./menus.html" class="active">Menus</a>
-                <a href="./orders.html">Order History</a>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <main>
-            <div class="admin-container">
-                <section class="stats">
-                    <div class="stat">
-                        <a href="#">
-                            <h2>
-                                2
-                                <i class="fa fa-bars"></i>
-                            </h2>
-                            <label>Menus</label>
-                        </a>
-                    </div>
-                    <div class="stat">
-                        <button id="newMenuBtn" class="btn btn-blue trigger">Set New Menu</button>
-                    </div>
-                </section>
-                <section>
-                    <h4>Available Menus</h4>
-                    <table id="dtable">
-                        <thead>
-                            <tr>
-                                <th>Menu Date</th>
-                                <th>Menu Type</th>
-                                <th>Meal Options</th>
-                                <th>Description</th>
-                                <th>Action</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-
-                            <tr>
-                                <td>04/05/2018</td>
-                                <td>Menu "Table d'hote </td>
-                                <td>1</td>
-                                <td>Lorem ipsum dolor sit amet consectetur adipisicing elit. Impedit dolor velit, dolorum debitis
-                                    rerum explicabo labore non, minus eum atque</td>
-                                <td>
-                                    <a href="./choose_options.html">
-                                        <button class="btn btn-g">Add Meal Options</button>
-                                    </a>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>04/04/2018</td>
-                                <td>Menu "du Jour".</td>
-                                <td>2</td>
-                                <td>Lorem ipsum dolor sit amet consectetur adipisicing elit. Impedit dolor velit, dolorum debitis
-                                    rerum explicabo labore non, minus eum atque</td>
-                                <td>
-                                    <a href="./choose_options.html">
-                                        <button class="btn btn-g">Add Meal Options</button>
-                                    </a>
-
-                                </td>
-                            </tr>
-                        </tbody>
-                    </table>
-                    <div class="modal">
-                        <div class="modal-content">
-                            <span class="close-button">
-                                <i class="fa fa-close"></i>
-                            </span>
-                            <h3>SET Menu</h3>
-                            <form method="POST">
-                                <div class="form-container">
-                                    <input class="form-input" placeholder="Menu Type" type="text" name="menu" autocomplete="off">
-                                    <input id="menuDate" placeholder="Menu Day" class="form-input" type="text" autocomplete="off" name="menuDate">
-                                    <textarea name="desc" id="" placeholder="Description" cols="30" rows="10"></textarea>
-                                    <input type="button" class="btn btn-blue" value="Add">
-                                </div>
-                            </form>
-                        </div>
-                    </div>
-                </section>
-            </div>
-        </main>
-        <footer>
-
-        </footer>
-        <script src="../js/jquery.min.js"></script>
-        <script src="../js/jquery-ui.min.js"></script>
-        <script src="../js/jquery.dataTables.min.js"></script>
-        <script src="../js/admin.js"></script>
-
-        <script>
-            $(document).ready(function () {
-                $("#dtable").DataTable({
-                    responsive: true,
-                    order: [
-                        [0, "desc"]
-                    ]
-                });
-                $("#menuDate").datepicker({
-                    dateFormat: "dd/mm/yy"
-                });
-
-                //trigger displaying and closing of a modal
-                var modal = document.querySelector(".modal");
-                var trigger = document.querySelector(".trigger");
-                var closeButton = document.querySelector(".close-button");
-
-                function toggleModal() {
-                    modal.classList.toggle("show-modal");
-                }
-
-                function windowOnClick(event) {
-                    if (event.target === modal) {
-                        toggleModal();
-                    }
-                }
-
-                trigger.addEventListener("click", toggleModal);
-                closeButton.addEventListener("click", toggleModal);
-                window.addEventListener("click", windowOnClick);
-            });
-        </script>
-    </body>
-
-</html>
