--- ui/admin/orders.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="../css/jquery-ui.min.css">
-        <link rel="stylesheet" href="../css/main.css">
-        <link rel="stylesheet" href="../css/admin.css">
-        <link rel="stylesheet" href="../css/jquery.dataTables.min.css">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css">
-        <title>Book A meal:Admin</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html">Admin</a>
-                <a href="./mealoptions.html">Meal Options</a>
-                <a href="./menus.html">Menus</a>
-                <a href="./orders.html" class="active">Order History</a>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <main>
-            <div class="admin-container">
-                <section>
-                    <h4>Order History</h4>
-                    <table id="dtable">
-                        <thead>
-                            <tr>
-                                <th>Customer's Name</th>
-                                <th>Customer's Contact</th>
-                                <th>Meal Options</th>
-                                <th>Total Cost</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr>
-                                <td>Nsubuga Solomon</td>
-                                <td>+256784019994 </td>
-                                <td>Beef with rice, Beef with fries</td>
-                                <td>UGX 40,500</td>
-                            </tr>
-                        </tbody>
-                    </table>
-
-                </section>
-            </div>
-        </main>
-        <footer>
-        </footer>
-        <script src="../js/jquery.min.js"></script>
-        <script src="../js/jquery.dataTables.min.js"></script>
-        <script src="../js/admin.js"></script>
-        <script>
-            $(document).ready(function () {
-                $("#dtable").DataTable({
-                    responsive: true,
-                    order: [
-                        [0, "desc"]
-                    ]
-                });
-
-            });
-        </script>
-    </body>
-
-</html>
