--- ui/caterer_signup.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="./css/login.css">
-        <title>Create Account</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html" class="active">Book A Meal</a>
-                <a href="./caterer_signup.html">Signup as a caterer</a>
-                <a href="./signup.html">Register</a>
-                <a style="right:0%" href="./login.html">Login</a>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <div class="title">
-            <h3>Create Account</h3>
-            <p>Sign up as a caterer</p>
-        </div>
-        <div class="container">
-            <div class="login-panel">
-                <form action="." method="POST">
-                    <div class="input-container">
-                        <div class="form-grp">
-                            <label for="email">Business Name</label>
-                            <input class="form-input" type="text" id="bname" name="businessName" autocomplete="off">
-                        </div>
-                        <div class="form-grp">
-                            <label for="email">Business Address</label>
-                            <input class="form-input" type="text" id="baddress" name="businessAddress" autocomplete="off">
-                        </div>
-                        <div class="form-grp">
-                            <label for="email">Full Name</label>
-                            <input class="form-input" type="text" id="fname" name="fullName" autocomplete="off">
-                        </div>
-                        <div class="form-grp">
-                            <label for="email">Email</label>
-                            <input class="form-input" type="email" id="email" name="email" autocomplete="off">
-                        </div>
-                        <div>
-                            <label for="password">Password</label>
-                            <input class="form-input" type="password" id="password" name="password" autocomplete="off">
-                        </div>
-                        <input class="btn btn-blue" type="button" value="Create Account">
-                        <p>already have an account?
-                            <a href="./login.html">Login</a>
-                        </p>
-                    </div>
-                </form>
-            </div>
-        </div>
-        <script>
-            function showMenu() {
-                var x = document.getElementById("myNavbar");
-                if (x.className === "navbar") {
-                    x.className += " responsive";
-                } else {
-                    x.className = "navbar";
-                }
-            }
-        </script>
-    </body>
-
-</html>
