--- ui/change_order.html
@@ -1,99 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="./css/main.css">
-        <link rel="stylesheet" href="./css/admin.css">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css">
-        <title>Book A meal</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html" class="active">Book A Meal</a>
-                <a style="right:0%" href="#">Logout</a>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <main>
-            <div class="admin-container">
-                <section class="stats">
-                    <div>
-                        <h4>Change your order.</h4>
-                    </div>
-                </section>
-                <section class="menu-items">
-                    <div class="meal-option">
-                        <div class="img-container">
-                            <img src="./img/food-dinner-pasta.jpg" alt="">
-                        </div>
-                        <div class="meal-info">
-                            <input class="checkbox" type="checkbox">
-                            <h4>Beef with fries</h4>
-                            <p>Cinamon flavour lorem ipsum</p>
-                        </div>
-                    </div>
-
-                    <div class="meal-option">
-                        <div class="img-container">
-                            <img src="./img/potatoes.jpeg" alt="">
-                        </div>
-                        <div class="meal-info">
-                            <input class="checkbox" type="checkbox">
-                            <h4>Potatoes and chicken</h4>
-                            <p>Potatoes with chicken stew flavour</p>
-                        </div>
-                    </div>
-                    <div class="meal-option">
-                        <div class="img-container">
-                            <img src="./img/photo3.jpeg" alt="">
-                        </div>
-                        <div class="meal-info">
-                            <input class="checkbox" type="checkbox">
-                            <h4>Cherry Tomatoes and Rice</h4>
-                            <p>English Coffee with cinamon flavour</p>
-                        </div>
-                    </div>
-                    <div class="meal-option">
-                        <div class="img-container">
-                            <img src="./img/burger.jpeg" alt="">
-                        </div>
-                        <div class="meal-info">
-                            <input class="checkbox" type="checkbox">
-                            <h4>Burger</h4>
-                            <p>Burgers mixed with lorem flavour</p>
-                        </div>
-                    </div>
-                </section>
-                <section class="order-total">
-                    <div>
-                        <label for="total">Total: </label>
-                        <label class="amt" for="amount">UGX 45,000</label>
-
-                        <button onclick="cancelOrder()" class="btn btn-d">Cancel</button>
-                        <button onclick="makeOrder()" class="btn btn-g">Proced to order</button>
-                    </div>
-                </section>
-            </div>
-        </main>
-        <footer>
-
-        </footer>
-        <script src="./js/jquery.min.js"></script>
-        <script src="./js/admin.js"></script>
-        <script>
-            function cancelOrder() {
-                window.document.location = './orders.html';
-            }
-
-            function makeOrder() {
-                window.document.location = './order.html';
-            }
-        </script>
-    </body>
-
-</html>
