--- ui/css/admin.css
@@ -1,119 +1,0 @@
-@import url("../css/common.css");
-@import url("../css/jquery.dataTables.min.css");
-
-.admin-container {
-  margin: 10px 20px 0px 20px;
-}
-
-.stats {
-  display: grid;
-  grid-gap: 10px;
-  margin-bottom: 10px;
-  grid-template-columns: repeat(auto-fit, minmax(100px, 1fr));
-}
-
-.stat {
-  display: grid;
-  align-items: center;
-  justify-content: center;
-  border: 1px solid #e8e9eb;
-  height: 100px;
-}
-
-.stat label {
-  font-size: 15px;
-  color: #ddd;
-  font-weight: bold;
-}
-
-table {
-  display: table;
-}
-
-td,
-th {
-  border: 1px solid #ddd;
-  padding: 8px;
-}
-#dtable tr:nth-child(even) {
-  background-color: #f2f2f2;
-}
-
-#dtable th {
-  padding-top: 12px;
-  padding-bottom: 12px;
-  text-align: left;
-  background-color: #00a1e0;
-  color: white;
-}
-
-.form-container {
-  margin-right: 10px;
-}
-.form-container textarea {
-  width: 100%;
-  border: 2px solid #ddd;
-  margin-bottom: 10px;
-}
-
-.form-container input {
-  margin-left: 0px;
-}
-
-.meals {
-  display: inline-block;
-}
-.meal-o {
-  display: inline-block;
-  margin-top: 10px;
-  margin-left: 10px;
-  border: 1px solid #e8e9eb;
-  width: 200px;
-  height: 300px;
-}
-
-.meal-actions {
-  display: inline-block;
-  margin: 10px;
-  justify-content: end;
-}
-
-.meal-actions input[type="checkbox"] {
-  width: 25px;
-  position: relative;
-}
-
-.meal-detail {
-  margin-left: 10px;
-  margin-right: 10px;
-}
-
-.thumbnail {
-  display: block;
-  margin-left: 10px;
-}
-
-.thumbnail img {
-  border-radius: 4px;
-  padding: 5px;
-  width: 150px;
-}
-
-.fa-close {
-  color: red;
-}
-
-.fa-edit {
-  color: #00a1e0;
-}
-
-@media screen and (max-width: 480px) {
-  .admin-container {
-    margin: 10px 20px 0px 20px;
-  }
-
-  .meals {
-    display: grid;
-    justify-content: center;
-  }
-}
