--- ui/css/common.css
@@ -1,151 +1,0 @@
-body {
-  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-}
-
-a {
-  color: #00a1e0;
-  text-decoration: none;
-}
-
-.header {
-  background-color: #333;
-}
-
-.navbar {
-  overflow: hidden;
-  margin: 0px 100px 0px 100px;
-}
-
-.navbar a {
-  float: left;
-  display: block;
-  color: #f2f2f2;
-  text-align: center;
-  padding: 14px 16px;
-  text-decoration: none;
-  font-size: 17px;
-}
-
-.navbar a:hover {
-  background-color: #e8e9eb;
-  color: black;
-}
-
-.active {
-  background-color: #00a1e0;
-  color: white;
-}
-
-.navbar .icon {
-  display: none;
-}
-
-.form-input {
-  display: block;
-  border: 1px solid #ced4da;
-  width: 100%;
-  border-radius: 0.25rem;
-  margin-bottom: 20px;
-  height: 35px;
-  font-size: 15px;
-}
-
-.btn {
-  height: 35px;
-  border: 1px solid #ced4da;
-  border-radius: 5px;
-  font-size: 15px;
-}
-
-.btn-blue {
-  background-color: #00a1e0;
-  border-color: #00a1e0;
-  color: #ffffff;
-}
-.btn-g {
-  background-color: rgb(0, 175, 67);
-  border-color: rgb(0, 175, 67);
-  color: #ffffff;
-}
-
-label {
-  display: inline-block;
-  margin-bottom: 0.5rem;
-}
-
-.float-right {
-  float: right;
-}
-
-.modal {
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  opacity: 0;
-  visibility: hidden;
-  transform: scale(1.1);
-  transition: visibility 0s linear 0.25s, opacity 0.25s 0s, transform 0.25s;
-}
-.modal-content {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: white;
-  padding: 1rem 1.5rem;
-  width: 24rem;
-  border-radius: 0.5rem;
-  border: 1px solid #f6f6f8;
-}
-.close-button {
-  float: right;
-  width: 1.5rem;
-  line-height: 1.5rem;
-  text-align: center;
-  cursor: pointer;
-  border-radius: 0.25rem;
-  background-color: lightgray;
-}
-.close-button:hover {
-  background-color: darkgray;
-}
-.show-modal {
-  opacity: 1;
-  visibility: visible;
-  transform: scale(1);
-  transition: visibility 0s linear 0s, opacity 0.25s 0s, transform 0.25s;
-}
-
-@media screen and (max-width: 480px) {
-  .modal-content {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background-color: white;
-    padding: 1rem 1.5rem;
-    width: 15rem;
-    border-radius: 0.5rem;
-    border: 1px solid #f6f6f8;
-    margin-top: 60px;
-    /* margin: 0px 50px 0px 50px; */
-  }
-}
-
-@media screen and (min-width: 500px) and (max-width: 600px) {
-  .modal-content {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background-color: white;
-    padding: 1rem 1.5rem;
-    width: 17rem;
-    border-radius: 0.5rem;
-    border: 1px solid #f6f6f8;
-    /* margin: 0px 50px 0px 50px; */
-  }
-}
