--- ui/css/login.css
@@ -1,36 +1,0 @@
-@import url("common.css");
-
-.title {
-  text-align: center;
-  padding: 50px 0 20px;
-}
-
-body {
-  background: #f6f6f8;
-}
-
-.title h3 {
-  color: #00a1e0;
-  font-size: 30px;
-}
-
-.title p {
-  font-weight: bold;
-}
-
-.login-panel {
-  width: auto;
-  height: auto;
-  border: 1px solid #ced4da;
-  background: #fff;
-}
-
-.input-container {
-  padding: 20px 20px 0px 20px;
-}
-
-.container {
-  display: grid;
-  justify-content: center;
-  align-items: center;
-}
