--- ui/index.html
@@ -1,132 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="./css/main.css">
-        <link rel="stylesheet" href="./css/jquery-ui.min.css">
-        <title>Book A meal</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html" class="active">Book A Meal</a>
-                <a style="right:0%" href="./orders.html"> My Orders</a>
-                <div class="float-right">
-                    <a href="./caterer_signup.html">Signup as a caterer</a>
-                    <a href="./signup.html">Register</a>
-                    <a style="right:0%" href="./login.html">Login</a>
-                </div>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <div class="app-banner">
-            <p>Everyday we present one featured menu</p>
-            <p>Allowing you to choose from the different meal options and order</p>
-            <img src="./img/plate.png" alt="logo">
-        </div>
-        <main>
-            <section class="specific-day">
-                <div>
-                    <h4>Check Out Today's Menu</h4>
-                    <p>With the Available Meal Options</p>
-                </div>
-                <div>
-                    <h4>Choose a specific day:</h4>
-                    <input id="menuDate" class="form-input" type="text" name="specificDay">
-                </div>
-            </section>
-
-            <section class="menu-items">
-                <div class="meal-option">
-                    <div class="img-container">
-                        <img src="./img/food-dinner-pasta.jpg" alt="">
-                    </div>
-                    <div class="meal-info">
-                        <input class="checkbox" type="checkbox">
-                        <label class="title">Beef with fries:</label>
-                        <label class="price" for="">UGX: 10,000</label>
-                        <p>Cinamon flavour lorem ipsum</p>
-                    </div>
-                </div>
-
-                <div class="meal-option">
-                    <div class="img-container">
-                        <img src="./img/potatoes.jpeg" alt="">
-                    </div>
-                    <div class="meal-info">
-                        <input class="checkbox" type="checkbox">
-                        <label class="title">Potatoes and chicken</label>
-                        <p class="price" for="">UGX: 20,000</p>
-                        <p>Potatoes with chicken stew flavour</p>
-                    </div>
-                </div>
-                <div class="meal-option">
-                    <div class="img-container">
-                        <img src="./img/photo3.jpeg" alt="">
-                    </div>
-                    <div class="meal-info">
-                        <input class="checkbox" type="checkbox">
-                        <label class="title">Cherry Tomatoes and Rice</label>
-                        <p class="price" for="">UGX: 12,000</p>
-                        <p>English Coffee with cinamon flavour</p>
-                    </div>
-                </div>
-                <div class="meal-option">
-                    <div class="img-container">
-                        <img src="./img/burger.jpeg" alt="">
-                    </div>
-                    <div class="meal-info">
-                        <input class="checkbox" type="checkbox">
-                        <label class="title">Burger</label>
-                        <p class="price" for="">UGX: 10,200</p>
-                        <p>Burgers mixed with lorem flavour</p>
-                    </div>
-                </div>
-            </section>
-            <section class="checkout">
-                <div class="checkout-img">
-                    <img src="./img/photo2.jpeg" alt="">
-                </div>
-                <div class="checkout-totals">
-                    <label for="total">Total:</label>
-                    <p for="">UGX 20000</p>
-                </div>
-                <div class="checkout-btn">
-                    <button onclick="proceedToCheckout()" class="btn btn-blue">Proceed to order</button>
-                </div>
-            </section>
-        </main>
-        <footer>
-
-        </footer>
-        <script src="./js/jquery.min.js"></script>
-        <script src="./js/jquery-ui.min.js"></script>
-
-        <script>
-            function showMenu() {
-                var x = document.getElementById("myNavbar");
-                if (x.className === "navbar") {
-                    x.className += " responsive";
-                } else {
-                    x.className = "navbar";
-                }
-            }
-
-            function proceedToCheckout() {
-                //localStorage.setItem("item", 24242);
-                window.document.location = './order.html';
-            }
-
-            $(document).ready(function () {
-                $("#menuDate").datepicker({
-                    dateFormat: 'dd/mm/yy'
-                });
-            })
-        </script>
-    </body>
-
-</html>
