--- ui/login.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <link rel="stylesheet" href="./css/login.css">
-        <title>Log In</title>
-    </head>
-
-    <body>
-        <header class="header">
-            <div class="navbar" id="myNavbar">
-                <a href="./index.html" class="active">Book A Meal</a>
-                <div class="float-right">
-                    <a href="./caterer_signup.html" class="active">Signup as a caterer</a>
-                    <a href="./signup.html">Register</a>
-                    <a style="right:0%" href="./login.html">Login</a>
-                </div>
-                <a href="javascript:void(0);" class="icon" onclick="showMenu()">&#9776;</a>
-            </div>
-        </header>
-        <div class="title">
-            <h3>Sign In</h3>
-        </div>
-        <div class="container">
-
-            <div class="login-panel">
-                <form>
-                    <div class="input-container">
-                        <div class="form-grp">
-                            <label for="email">Email</label>
-                            <input required class="form-input" type="email" id="email" name="email">
-                        </div>
-                        <div>
-                            <label for="password">Password</label>
-                            <input required class="form-input" type="password" id="password" name="password">
-                        </div>
-                        <input onclick="login()" type="button" class="btn btn-blue" value="Login">
-                        <p>Don't have an account?
-                            <a href="./signup.html">Sign up</a>
-                        </p>
-                    </div>
-                </form>
-            </div>
-        </div>
-        <script>
-            function showMenu() {
-                var x = document.getElementById("myNavbar");
-                if (x.className === "navbar") {
-                    x.className += " responsive";
-                } else {
-                    x.className = "navbar";
-                }
-            }
-
-            function login() {
-                window.document.location = './admin/index.html';
-            }
-        </script>
-    </body>
-
-</html>
